Bank Payment Gateway
This a payment gateway system for banks implemented in Python using openpyxl and pandas library. This payment gateway system facilitates the current customers of bank to securely login through password and email Id with OTP verification via mail Id. It provides the user with operations such as debit,  credit and print bank statement.